Enable texture binding and fix parameter type
Uncommented the line to set the texture in AppWindow::onUpdate, enabling texture binding for rendering. Also changed the type of size_byte_shader from UINT to size_t in VertexBuffer.cpp for better type safety.

+++ GDENG03 Getting Started Game Engine/VertexBuffer.cpp
@@ -7,7 +7,7 @@ VertexBuffer::VertexBuffer(RenderSystem* system, void* list_vertices, UINT size_
 	// can be validated against the array of attributes just created
 	//...To be discussed further later on
 	void* shader_byte_code, 
-	UINT size_byte_shader) : m_render_system(system)
+	size_t size_byte_shader) : m_render_system(system)
 	,m_buffer(0), m_layout(0)
 {
 	/*
